Adding doublets & starting a ramp
doublet.lua: Modified to incorporate doublets from ailerons and throttle inputs.

ramp.lua: Added to eventually replace doublet.lua. This code includes the option to run a ramp instead of a step. The intent of this is mainly for throttle testing. This code is NOT in a runnable state currently.